requirements.txt
googledrivedownloader
schedule
psycopg2-binary
pyTelegramBotAPI